refactor: extract recommendations hook module
